Implement a bubble_sort method
Add the bubble_sort method which compare two consecutive slots
and order them, repeating the process until the end of the array
and then repeat the process array.lengt-1 number of times.

This method follows every project's requirement, it follows the
bubble sorting process but optimizes the repeating steps to don't
compare the ordered sub array at the end.